Bare minimum changes to get the UI hooked up properly
