+++ Assets/Shopify/UIToolkit/Shops/Generic/Classes/ProductDetailsViewBindings.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 1e2f0bd79f1504d929b109daecc79d10
+timeCreated: 1516731902
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
