+++ Assets/Shopify/UIToolkit/Shops/Generic/Classes/ProductListView.cs
@@ -1,6 +1,7 @@
 ﻿namespace Shopify.UIToolkit.Shops.Generic {
     using System.Collections;
     using System.Collections.Generic;
+    using System.Linq;
     using Shopify.Unity;
     using UnityEngine;
     using UnityEngine.UI;
@@ -15,6 +16,9 @@
                 listItem.gameObject.SetActive(true);
                 listItem.transform.SetParent(_scrollContentRect.transform, false);
                 listItem.SetProduct(product);
+                listItem.OnClick.AddListener(() => {
+                    Shop.ViewProductDetails(product, product.variants().edges().Select((x) => x.node()).ToArray());
+                });
             }
         }
